DataLib Hdf5 查询 写入初步成型

+++ src/SQLiteLib/Shared/DataLib.Shared/Table/Impl/DataRow.cs
@@ -17,7 +17,7 @@ namespace DataLib.Table.Impl
         /// <summary>
         /// Row Index
         /// </summary>
-        public int RowIndex { get; set; }
+        public ulong RowIndex { get; set; }
 
         /// <summary>
         /// 数据结果集合
